refactor(bridge-status-controller): colocate failure classifiers in metrics utils
Fold submit/status failure telemetry helpers into the existing metrics files instead of a new module.
